requirements.txt
pandas==2.2.3
numpy==1.26.0
nltk==3.9.1
textblob==0.19.0
gensim==4.3.3
textstat==0.7.5
tensorflow-cpu==2.18.0
matplotlib==3.10.0
scikit-learn==1.6.1
scipy==1.13.1
Flask==2.2.5
syllapy==0.7.2
xgboost==2.1.4

# API
fastapi==0.115.11         # API framework
uvicorn==0.34.0        # web server
python-multipart==0.0.20


# tests
httpx==0.28.1          # HTTP client
pytest-asyncio==0.25.3 # asynchronous I/O support for pytest
